add show & search bookingServices + update logic calculate reward points for customers and update customerType

+++ src/main/java/com/lullabyhomestay/homestay_management/domain/Customer.java
@@ -47,7 +47,7 @@ public class Customer {
     private String password;
 
     @Column(name = "RewardPoints")
-    private float rewardPoints;
+    private Double rewardPoints;
 
     @Column(name = "IsLocked")
     private boolean isLocked;
